Global update
It was hard, I look out the window and it beckons me... In general, I changed the architecture, made the code much cleaner, and you can say it looks more like a "Clean Architecture". But it's still not perfect, I'll add new technologies in the future and make it even cleaner.