global update 2.0
I'm very tired boss... I added jwt registration and added a protective layer for important methods, and if I'm not mistaken, I fixed a couple of past errors and bugs, as well as config and its handler. But there are bugs and minor mistakes that I will definitely fix! I also have a lot of ideas for improving the entire architecture, which I will do very soon. FULL FOCUS